requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
allure-python-commons-test==2.8.18
requests
pytest==6.0.1
pytest-forked==1.3.0
pytest-pycharm==0.7.0
pytest-timeout==1.4.2
pytest-xdist==2.0.0
